Code Update #2

+++ src/main/java/com/teenkung/craftycrates/utils/RarityStorage.java
@@ -3,16 +3,16 @@ package com.teenkung.craftycrates.utils;
 public class RarityStorage {
 
     private String display;
-    private Double chance;
+    private Integer chance;
     private String pool;
-    public RarityStorage(String display, Double chance, String pool) {
+    public RarityStorage(String display, Integer chance, String pool) {
         this.display = display;
         this.chance = chance;
         this.pool = pool;
     }
 
     public String getDisplay() { return display; }
-    public Double getChance() { return chance; }
+    public Integer getChance() { return chance; }
     public String getPool() { return pool; }
 
 }
